refactor: paht를 이용하여 pathResponse를 반환하도록 수정
